refactor: api 요청을 post 메서드로 분리
사물함 반납, 등록, 연장 API 요청을 PUT 메서드에서 POST 메서드로 분리하여 API의 역할과 책임을 명확히 합니다.
또한, v2 API 응답 타입 정의를 추가하여 API 응답 구조를 명확하게 정의합니다.
