+++ src/entities/locker/api/get.ts
@@ -1,11 +1,5 @@
 import { API } from '@/shared';
 
-type ApiResponse<T> = {
-  code: string;
-  message: string;
-  data: T;
-};
-
 export const getLockerLocations = async () => {
   const { data } = await API.get<Locker.LockerLocationsResponseDto>(
     '/api/v1/lockers/locations',
@@ -22,26 +16,26 @@ export const getLockerList = async (locationId: string) => {
 
 export const getMyLockerV2 = async () => {
   const { data } =
-    await API.get<ApiResponse<LockerV2.MeResponse>>('/api/v2/lockers/me');
+    await API.get<Api.V2Response<LockerV2.MeResponse>>('/api/v2/lockers/me');
   return data.data;
 };
 
 export const getLockerLocationV2 = async (locationId: string) => {
-  const { data } = await API.get<ApiResponse<LockerV2.LocationDetail>>(
+  const { data } = await API.get<Api.V2Response<LockerV2.LocationDetail>>(
     `/api/v2/lockers/locations/${locationId}`,
   );
   return data.data;
 };
 
 export const getLockerLocationsV2 = async () => {
-  const { data } = await API.get<ApiResponse<LockerV2.LocationsOverview>>(
+  const { data } = await API.get<Api.V2Response<LockerV2.LocationsOverview>>(
     '/api/v2/lockers/locations',
   );
   return data.data;
 };
 
 export const getLockerPeriodStatusV2 = async () => {
-  const { data } = await API.get<ApiResponse<LockerV2.PeriodStatus>>(
+  const { data } = await API.get<Api.V2Response<LockerV2.PeriodStatus>>(
     '/api/v2/lockers/period-status',
   );
   return data.data;
