+++ src/entities/locker/api/index.ts
@@ -1,2 +1,3 @@
 export * from './get';
 export * from './put';
+export * from './post';
